auth: support threebot token validation
Since using threebot token on the hub, the hub can generate a token
which is validated with it's own signature. The easiest way to ensure
a token is valid is to validate it against the hub itself.

This update detect if token is a jwt or not and validate the token with
the right method.

This fixes #6